Navigation routes
Added a navBar with react-router, and can now navigate through components as well as a 404 and a home page. Added conditional rendering so i wont get an error if a country doesn't have certain Info. Pretty close to being functionally done just need a way to get the graphs going as well as a way to conditionally render a grid based on length of countries array on finder page. Maybe want to add context on state so it persists the same base country for all pages