Cards ,Conditons & Graphs
Added a bit of styling throughout Project. Flex wrap on card grid.  Added conditional rendering to improve UX as well as to have it not crash on initial renders.  Lots of code cleanup

Starting on graph api and migrated some functions over to it. Got graph component to be aware of state just need it to render.
Added a bit of styling throughout Project. Flex wrap on card grid.  Added conditional rendering to improve UX as well as to have it not crash on initial renders.  Lots of code cleanup

Starting on graph api and migrated some functions over to it. Got graph component to be aware of state just need it to render.