Date Presets
TRYING to get date presets working. Proving to be a bit harder. Currently have a function that can go back in time based upon preset time. For some reason the array dosent want to build from inbetween the two dates. Also have a funtion that plucs based on the preset whether its every other element of every fifth element. So just need the array to build; then need it to plot on the graph and have the labels update